fix: potentially printing wrong error message
it's possible for the close() calls to override the errno resulting in
incorrect error message being printed.

call error() immediately or save the `errno`s to tmp variables to avoid
such possibilities.

additionally, pfd_{read,write} doesn't need to be initialized to -1.
that was needed back from when an earlier implementation (which never
made it into master) was using a custom `close` wrapper which would
no-op in case of negative fds.