code-style: reduce noise by using a pointer
the continuous appearance of `img->multi.` clutters the code quite a
lot. using a pointer to make the expression shorter and thus reduce
noise makes sense.

this technique isn't new to the code-base, for example see the usage of
pointer `e` inside the function `win_init()`.

this could be applied to the webp and gif loader with relative ease,
since it's a matter of search and replace `s/img->multi\./m->/g`.

but given those loaders will be removed in the future, probably best to
leave them be.