img_render: explicitly set blending
this was the cause of the transparency bug for animated pngs.
img_load_multiframe() was messing with the blend flag and thus it messed
with the rendering function.

the fix is to explicitly enable blending inside the rendering function.

also explicitly set operation as well just in case some other place in
the codebase messes with it in the future.

Closes: https://codeberg.org/nsxiv/nsxiv/issues/300